[IMP] separate files per model, move action_assign method to sale_line_quant_extended
