+++ sale_line_quant_extended/models/__init__.py
@@ -3,5 +3,7 @@
 from . import sale_order
 from . import sale_order_line
 from . import procurement
-from . import stock
+from . import stock_quant
+from . import stock_move
+from . import stock_picking
 from . import purchase
